requirements.txt

apache-airflow==2.2.4
cronitor==4.4.4
